Refactor notification handling by introducing NotificationState and updating message input and pages to use the new state management
