+++ ui/lib/main.dart
@@ -7,11 +7,15 @@ import 'package:provider/provider.dart';
 // Project imports:
 import 'package:ui/app_state.dart';
 import 'package:ui/pages/home_page.dart';
+import 'package:ui/state/notification_state.dart';
 
 void main() {
   runApp(
-    ChangeNotifierProvider(
-      create: (context) => AppState(),
+    MultiProvider(
+      providers: [
+        ChangeNotifierProvider(create: (context) => AppState()),
+        ChangeNotifierProvider(create: (context) => NotificationState()),
+      ],
       child: const AIApp(),
     ),
   );
