+++ ui/lib/pages/login_page.dart
@@ -8,6 +8,7 @@ import 'package:provider/provider.dart';
 // Project imports:
 import 'package:ui/app_state.dart';
 import 'package:ui/helpers/http_helper.dart';
+import 'package:ui/state/notification_state.dart';
 
 class LoginPage extends StatefulWidget {
   final HttpHelper httpHelper;
@@ -37,6 +38,7 @@ class _LoginPageState extends State<LoginPage> {
   @override
   Widget build(BuildContext context) {
     final appState = Provider.of<AppState>(context, listen: false);
+    final notificationState = Provider.of<NotificationState>(context);
 
     Widget ipTextField() {
       return TextField(
@@ -94,10 +96,10 @@ class _LoginPageState extends State<LoginPage> {
               appState.clearMessages();
               appState.addMessage(message);
               appState.setActivePage('message');
-              appState.clearNotification();
+              notificationState.clearNotification();
             }
           } catch (e) {
-            appState.setNotificationError('Failed to connect: $e');
+            notificationState.setNotificationError('Failed to connect: $e');
           }
         },
       );
